Open SlateDB instance only once on startup
Previously the database was opened twice during initialization:
once unencrypted for key management, then closed and reopened
with encryption. Now we build a single SlateDB instance, use it
for key management, then wrap it with encryption.
